requirements.txt
autopep8==1.5.7
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
et-xmlfile==1.1.0
idna==2.10
lxml==4.6.3
openpyxl==3.0.7
pycodestyle==2.7.0
pygame==2.0.1
pytelebot==0.0.1
requests==2.25.1
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.5
